yolnpCVA/requirements.txt
absl-py==1.3.0
astor==0.8.1
asttokens==2.2.0
attrdict==2.0.1
attrs==22.1.0
Babel==2.11.0
backcall==0.2.0
bce-python-sdk==0.8.74
beautifulsoup4==4.11.1
cachetools==5.2.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.6
cssselect==1.2.0
cssutils==2.6.0
cycler==0.11.0
Cython==0.29.32
decorator==5.1.1
dill==0.3.6
dnspython==2.2.1
et-xmlfile==1.1.0
exceptiongroup==1.0.4
executing==1.2.0
filterpy==1.4.5
fire==0.4.0
Flask==2.2.2
Flask-Babel==2.0.0
Flask-Cors==3.0.10
fonttools==4.38.0
future==0.18.2
google-auth==2.15.0
google-auth-oauthlib==0.4.6
grpcio==1.51.1
idna==3.4
imageio==2.22.4
imgaug==0.4.0
importlib-metadata==5.0.0
iniconfig==1.1.1
Inject==4.3.1
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
kiwisolver==1.4.4
lanms-neo==1.0.2
lmdb==1.3.0
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
marshmallow==3.19.0
marshmallow-dataclass==8.5.10
matplotlib==3.5.3
matplotlib-inline==0.1.6
multiprocess==0.70.14
mypy-extensions==0.4.3
networkx==2.6.3
numpy==1.21.6
oauthlib==3.2.2
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
openpyxl==3.0.10
opt-einsum==3.3.0
packaging==21.3
paddle-bfloat==0.1.7
paddleocr==2.6.1.0
paddlepaddle==2.4.0
pandas==1.3.5
parso==0.8.3
pdf2docx==0.5.6
pickleshare==0.7.5
Pillow==9.3.0
pluggy==1.0.0
premailer==3.10.0
prompt-toolkit==3.0.33
protobuf==3.20.0
psutil==5.9.4
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyclipper==1.3.0.post4
pycryptodome==3.15.0
Pygments==2.13.0
pymongo==4.3.3
PyMuPDF<1.21.0
pyparsing==3.0.9
pytesseract==0.3.10
pytest==7.2.0
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==0.21.0
pytz==2022.6
PyWavelets==1.4.0
PyYAML==6.0
rapidfuzz==2.13.2
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
schedule==1.1.0
scikit-image==0.19.3
scipy==1.7.3
Polygon3
seaborn==0.12.1
Shapely==1.8.5.post1
six==1.16.0
soupsieve==2.3.2.post1
stack-data==0.6.2
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==2.1.0
thop==0.1.1.post2209072238
tifffile==2021.11.2
tomli==2.0.1
torch==1.13.0
torchvision==0.14.0
tqdm==4.64.1
traitlets==5.6.0
typing-inspect==0.8.0
typing_extensions==4.4.0
urllib3==1.26.12
visualdl==2.4.1
wcwidth==0.2.5
Werkzeug==2.2.2
wget==3.2
wincertstore==0.2
zipp==3.10.0
